RGB LEDS
ESP32 (wemos) provides an specific peripheral to generate pwm signals. It was primarily designed to control the intensity of LEDs, although it can also be used to generate PWM signals for other purposes. In our case, we will also use it to set the speed of the fan (DC motor).  It has 16 channels which can generate independent waveforms, 

In order to use this peripheral hardware with the wemos board we need to use ledc library. We will specifically use the following functions: